Adding event now adds a category, working create/delete review

+++ src/main/java/com/hiutaleapp/controller/ReviewController.java
@@ -1,13 +1,26 @@
 package com.hiutaleapp.controller;
 
 import com.hiutaleapp.dto.ReviewDTO;
+import com.hiutaleapp.entity.Event;
 import com.hiutaleapp.entity.Review;
+import com.hiutaleapp.entity.User;
 import com.hiutaleapp.service.ReviewService;
+import com.hiutaleapp.util.DataViolationException;
+import com.hiutaleapp.util.DatabaseConnectionException;
+import com.hiutaleapp.util.NotFoundException;
+import com.hiutaleapp.util.ReviewForm;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessResourceFailureException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authorization.AuthorizationDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.transaction.CannotCreateTransactionException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/reviews")
@@ -28,8 +41,29 @@ public class ReviewController {
     }
 
     @PostMapping("/create")
-    public ReviewDTO createReview(@RequestBody Review review) {
-        return reviewService.createReview(review);
+    public ReviewDTO createReview(@RequestBody ReviewForm reviewForm) {
+        if (reviewForm.getRating() > 5 || reviewForm.getRating() < 0) throw new DataViolationException("Rating must be between 5 and 0");
+        try {
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+            Review review = new Review();
+            User user = new User();
+            Event event = new Event();
+
+            user.setUserId(Long.parseLong(auth.getName()));
+            event.setEventId(reviewForm.getEventId());
+
+            review.setUser(user);
+            review.setEvent(event);
+            review.setRating(reviewForm.getRating());
+            review.setReviewText(reviewForm.getText());
+
+            return reviewService.createReview(review);
+        } catch (CannotCreateTransactionException e) {
+            throw new DatabaseConnectionException("Could not connect to the database");
+        } catch (DataIntegrityViolationException e) {
+            throw new DataViolationException("Could not create review due to nonexistent event key");
+        }
     }
 
     @PutMapping("/update/{id}")
@@ -39,6 +73,22 @@ public class ReviewController {
 
     @DeleteMapping("/delete/{id}")
     public void deleteReview(@PathVariable Long id) {
-        reviewService.deleteReview(id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        try {
+            Optional<ReviewDTO> review = reviewService.getReviewById(id);
+            if (review.isPresent()) {
+                if (review.get().getUserId() == Long.parseLong(auth.getName())) {
+                    reviewService.deleteReview(id);
+                } else {
+                    throw new AuthorizationDeniedException("You do not have permission to delete this review");
+                }
+            } else {
+                throw new NotFoundException("Review with this ID does not exist");
+            }
+        } catch (DataAccessResourceFailureException e) {
+            throw new DatabaseConnectionException("Could not connect to the database");
+        } catch (DataIntegrityViolationException e) {
+            throw new DataViolationException("Could not delete this review because other data is dependent upon it");
+        }
     }
 }
